feat: add hook for interaction with local storage
