+++ task-1/src/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useCharactesClose';
+export * from './useSearchQuery';
