+++ task-1/src/views/search-page/utils/index.ts
@@ -1,2 +1,1 @@
-export * from './terms';
 export * from './input';
