--- task-1/src/views/search-page/utils/terms/index.ts
@@ -1,2 +1,0 @@
-export * from './terms-from-string';
-export * from './terms-to-string';
