--- task-1/src/views/search-page/utils/terms/terms-from-string.ts
@@ -1,13 +1,0 @@
-export function getTermsFromString(
-  stroke: string | null
-): string[] {
-  if (stroke === '' || stroke === null) {
-    return [];
-  } else {
-    return (
-      stroke
-        ?.split(/(?<=") (?=")/)
-        .map((s) => s.slice(1, s.length - 1)) ?? []
-    );
-  }
-}
